Overhauled the userscript API help topics! macro
